modulesync 5.2.0

+++ spec/support/acceptance/prepare_host.rb
@@ -1,18 +1,16 @@
+# frozen_string_literal: true
+
 def prepare_host
   if fact('os.family') == 'RedHat'
     shell('rm -rf /etc/yum.repos.d/Zabbix*.repo; rm -rf /var/cache/yum/x86_64/*/Zabbix*; yum clean all --verbose')
     # The CentOS docker image has a yum config that won't install docs, to keep used space low
     # zabbix packages their SQL file as doc, we need that for bootstrapping the database
-    if fact('os.release.major').to_i == 7
-      shell('sed -i "/nodocs/d" /etc/yum.conf')
-    end
+    shell('sed -i "/nodocs/d" /etc/yum.conf') if fact('os.release.major').to_i == 7
   end
 
   # The Ubuntu 18.04+ docker image has a dpkg config that won't install docs, to keep used space low
   # zabbix packages their SQL file as doc, we need that for bootstrapping the database
-  if fact('os.distro.id') == 'Ubuntu'
-    shell('rm -f /etc/dpkg/dpkg.cfg.d/excludes')
-  end
+  shell('rm -f /etc/dpkg/dpkg.cfg.d/excludes') if fact('os.distro.id') == 'Ubuntu'
 
   # On Debian it seems that make is searching for mkdir in /usr/bin/ but mkdir
   # does not exist. Symlink it from /bin/mkdir to make it work.
